Implements restricted access to DeleteProduct

+++ SimpleStore.ConsoleUI/MenusLogic/ManagerLogic.cs
@@ -1,4 +1,5 @@
-﻿using SimpleStore.Domain.Manager.ManagerOperations;
+﻿using SimpleStore.Domain.Manager.ManagerLogin;
+using SimpleStore.Domain.Manager.ManagerOperations;
 using SimpleStore.Domain.Products;
 using System;
 using System.Collections.Generic;
@@ -92,6 +93,11 @@ namespace SimpleStore.ConsoleUI.MenusLogic
 
         public bool DeleteProduct(List<string> inputs)
         {
+            if (ManagerLogger.CurrentManager.ManagerPermission.PermissionTitle != "Super Admin")
+            {
+                throw new Exception("Only Super Admin is allowed");
+            }
+
             bool isDeletionSuccess = _productsOperator.DeleteProduct(inputs[0]);
             if (isDeletionSuccess == true)
             {
